requirements.txt
langchain-community
langchain-text-splitters
groq
sentence-transformers
faiss-cpu
numpy
json
pydantic
fastapi
uvicorn 